# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zenith Backend API",
    "description": "Paper-trading API for the Zenith options protocol.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "crate::alerts"
        ],
        "operationId": "get_alerts",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alert"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "crate::alerts"
        ],
        "operationId": "create_alert",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alert"
                }
              }
            }
          },
          "400": {
            "description": "Invalid alert",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown underlying",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/positions/open": {
      "post": {
        "tags": [
          "crate::positions"
        ],
        "operationId": "open_position",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenPositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              }
            }
          },
          "400": {
            "description": "Invalid position",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Insufficient buying power",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/positions/{id}/close": {
      "post": {
        "tags": [
          "crate::positions"
        ],
        "operationId": "close_position",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Position identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Position not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/positions/{id}/roll": {
      "post": {
        "tags": [
          "crate::positions"
        ],
        "summary": "Closes the given position and immediately opens its replacement (same\nunderlying/option_type/position_type/contracts, new strike and expiry)\nas one atomic transaction — either both happen or neither does.",
        "operationId": "roll_position",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Position identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollPositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid replacement position",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Position not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/strategies/execute": {
      "post": {
        "tags": [
          "crate::strategies"
        ],
        "summary": "Opens every leg of a multi-leg strategy (straddle, spread, iron\ncondor, ...) as one atomic transaction under a shared strategy_id —\neither all legs open or none do, so a mid-strategy insufficient-funds\nrejection can't leave a naked partial position behind.",
        "operationId": "execute_strategy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteStrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Position"
                  }
                }
              }
            }
          },
          "400": {
            "description": "A strategy must contain at least two legs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Insufficient buying power",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Alert": {
        "type": "object",
        "required": [
          "id",
          "wallet_address",
          "underlying",
          "condition",
          "target_price",
          "triggered",
          "created_at"
        ],
        "properties": {
          "condition": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "target_price": {
            "type": "number",
            "format": "double"
          },
          "triggered": {
            "type": "boolean"
          },
          "triggered_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "underlying": {
            "type": "string"
          },
          "wallet_address": {
            "type": "string"
          }
        }
      },
...
      "CreateAlertRequest": {
        "type": "object",
        "required": [
          "underlying",
          "condition",
          "target_price"
        ],
        "properties": {
          "condition": {
            "type": "string"
          },
          "target_price": {
            "type": "number",
            "format": "double"
          },
          "underlying": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "OpenPositionRequest": {
        "type": "object",
        "required": [
          "underlying",
          "strike",
          "expiry_days",
          "option_type",
          "position_type",
          "contracts"
        ],
        "properties": {
          "contracts": {
            "type": "number",
            "format": "double"
          },
          "expiry_days": {
            "type": "number",
            "format": "double"
          },
          "option_type": {
            "type": "string"
          },
          "position_type": {
            "type": "string"
          },
          "strike": {
            "type": "number",
            "format": "double"
          },
          "underlying": {
            "type": "string"
          }
        }
      },
...
      "RollPositionRequest": {
        "type": "object",
        "required": [
          "new_strike",
          "new_expiry_days"
        ],
        "properties": {
          "new_expiry_days": {
            "type": "number",
            "format": "double"
          },
          "new_strike": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "RollResult": {
        "type": "object",
        "required": [
          "closed",
          "opened"
        ],
        "properties": {
          "closed": {
            "$ref": "#/components/schemas/Position"
          },
          "opened": {
            "$ref": "#/components/schemas/Position"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque session token"
      }
    }
  }
}